Moved service tests to its proper test class, fixed so they work except for one
